Link to fix for Quill is defined error in modules
When trying to use a Quill module from Nuxt/SSR that itself uses Webpack, you must define Quill and window.Quill as Webpack plugins for the module to be able to find Quill. The linked to issue contains instructions and a fix. It also shows a different way to include Quill without SSR in a Nuxt app that allows use of the quill-editor component instead of a div.