replace the usage of pasteHTML() with setContents()
pasteHTML() causes the browser to focus on the editor (since this imitates the user pasting text, the user would want to focus on the editor that they are pasting text inside of). If the editor is offscreen when this is done, the browser scrolls to the editor to focus it. 

Since the value of v-model is often changed when the page is first loaded, if there are multiple vue-quill-editors on the page, it can cause the page to become jittery (while the focus is jumping around on each of the v-model updates, because of the pasteHTML() call).